refactor: :recycle: create formEditRegister component to be used in both pages
